+++ IbFlexReader/IbFlexReader.Contracts/CFDCharges.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class CFDCharges
+    {
+        public List<CFDCharge> CFDCharge { get; set; }
+    }
+}
