+++ IbFlexReader/IbFlexReader.Contracts/CashTransactions.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class CashTransactions
+    {
+        public List<CashTransaction> CashTransaction { get; set; }
+    }
+}
