+++ IbFlexReader/IbFlexReader.Contracts/ChangeInDividendAccruals.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class ChangeInDividendAccruals
+    {
+        public List<ChangeInDividendAccrual> ChangeInDividendAccrual { get; set; }
+    }
+}
