+++ IbFlexReader/IbFlexReader.Contracts/ConversionRates.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class ConversionRates
+    {
+        public List<ConversionRate> ConversionRate { get; set; }
+    }
+}
