+++ IbFlexReader/IbFlexReader.Contracts/EquitySummaryInBase.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class EquitySummaryInBase
+    {
+        public List<EquitySummaryByReportDateInBase> EquitySummaryByReportDateInBase { get; set; }
+    }
+}
