+++ IbFlexReader/IbFlexReader.Contracts/FlexStatements.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    public class FlexStatements
+    {
+        public FlexStatement FlexStatement { get; set; }
+
+        public int? Count { get; set; }
+    }
+}
