+++ IbFlexReader/IbFlexReader.Contracts/IbFlexReader.Contracts.csproj
@@ -17,6 +17,10 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
+  </ItemGroup> 
+	
+  <ItemGroup>
+    <Folder Include="Ib\" />
   </ItemGroup>
 
 </Project>
