+++ IbFlexReader/IbFlexReader.Contracts/Lot.cs
@@ -1,0 +1,6 @@
+﻿namespace IbFlexReader.Contracts
+{
+    public class Lot : Trade
+    {
+    }
+}
