+++ IbFlexReader/IbFlexReader.Contracts/OpenDividendAccruals.cs
@@ -1,0 +1,7 @@
+﻿namespace IbFlexReader.Contracts
+{
+    public class OpenDividendAccruals
+    {
+        public OpenDividendAccrual OpenDividendAccrual { get; set; }
+    }
+}
