+++ IbFlexReader/IbFlexReader.Contracts/OpenPositions.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class OpenPositions
+    {
+        public List<OpenPosition> OpenPosition { get; set; }
+    }
+}
