+++ IbFlexReader/IbFlexReader.Contracts/PriorPeriodPositions.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class PriorPeriodPositions
+    {
+        public List<PriorPeriodPosition> PriorPeriodPosition { get; set; }
+    }
+}
