+++ IbFlexReader/IbFlexReader.Contracts/SecuritiesInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class SecuritiesInfo
+    {
+        public List<SecurityInfo> SecurityInfo { get; set; }
+    }
+}
