+++ IbFlexReader/IbFlexReader.Contracts/SecurityInfo.cs
@@ -1,0 +1,53 @@
+﻿namespace IbFlexReader.Contracts
+{
+    public class SecurityInfo
+    {
+        public string AssetCategory { get; set; }
+
+        public string Symbol { get; set; }
+
+        public string Description { get; set; }
+
+        public string Conid { get; set; }
+
+        public string SecurityID { get; set; }
+
+        public string SecurityIDType { get; set; }
+
+        public string Cusip { get; set; }
+
+        public string Isin { get; set; }
+
+        public string ListingExchange { get; set; }
+
+        public string UnderlyingConid { get; set; }
+
+        public string UnderlyingSymbol { get; set; }
+
+        public string UnderlyingSecurityID { get; set; }
+
+        public string UnderlyingListingExchange { get; set; }
+
+        public string Issuer { get; set; }
+
+        public string Multiplier { get; set; }
+
+        public string Strike { get; set; }
+
+        public string Expiry { get; set; }
+
+        public string PutCall { get; set; }
+
+        public string PrincipalAdjustFactor { get; set; }
+
+        public string Maturity { get; set; }
+
+        public string IssueDate { get; set; }
+
+        public string UnderlyingCategory { get; set; }
+
+        public string SubCategory { get; set; }
+
+        public string Code { get; set; }
+    }
+}
