+++ IbFlexReader/IbFlexReader.Contracts/TradeConfirms.cs
@@ -1,0 +1,9 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class TradeConfirms
+    {
+        public List<TradeConfirm> TradeConfirm { get; set; }
+    }
+}
