+++ IbFlexReader/IbFlexReader.Contracts/Trades.cs
@@ -1,0 +1,10 @@
+﻿namespace IbFlexReader.Contracts
+{
+    using System.Collections.Generic;
+
+    public class Trades
+    {
+        public List<Lot> Lot { get; set; }
+        public List<Trade> Trade { get; set; }
+    }
+}
